🐛 fix &  ♻️ refactor : Fix type error & Reorganize folder structure of list page
