+++ src/app/list/[slug]/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import FormPageContainer from '@/components/FormPage/Container/FormPageContainer';
+import FormPageContainer from '@/app/list/[slug]/_components/Formpage/Container/FormPageContainer';
 import { ListProps } from '@/interfaces/listPage';
 import { ListItemDetail } from '@/types/apis/listItem';
