+++ src/app/list/[slug]/_components/Formpage/utils/getCategoryTypeFromSlug.ts
