+++ src/app/list/[slug]/loading.tsx
@@ -1,0 +1,3 @@
+export default function loading() {
+  return <div>loading</div>;
+}
